feat(terminal): trigger TermOSC autocommand events
This commit implements a new TermOSC autocommand event and has Neovim
emit this event when children of terminal buffers emit an OSC sequence
libvterm does not handle.

The TermOSC autocommand event has additional data in its v:event :
`command`, which corresponds to the OSC number (e.g. `7` in `OSC7`) and
`payload`, which is the actual content of the command.

This commit is meant as a first step in the direction of closing #2252.